requirements.txt
fastapi
uvicorn
pillow
black
pytest
httpx
python-multipart
